+++ ui-design-lib/src/main/java/co/electriccoin/zcash/ui/design/util/LightOrDark.kt
@@ -1,0 +1,9 @@
+package co.electriccoin.zcash.ui.design.util
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.ReadOnlyComposable
+
+@Composable
+@ReadOnlyComposable
+infix fun <T> T.orDark(dark: T): T = if (isSystemInDarkTheme()) dark else this